temp_check/requirements.txt
dify-plugin-sdk
requests
